Active reading support added.
Added a document that describes active reading for BA4ABW. At some
level, this is better than the CompOrg version... or, not. Actually,
this is just more appropriate to a 100-level course. It is, essentially,
the same thing as the CompOrg process, however.

There may be some pain in the Active Reading Support Environment, but
it is supposed to be for the students' own good.